Build and utilize pre-compiled binaries
Making users wait for a `stack install weeder` is an unnecessary tax on
every CI run that lacks a complete cache. Instead, this PR adds an
"Binaries" action that builds binaries across a runner/GHC matrix and
stores them on a Release for use by the action. The Binaries action is
idempotent and has been run and added all the binaries already, which
our own CI uses successfully.

I made binaries for Linux and Mac and GHCs 8.8.1 to 9.2.5. Windows may
be possible, but the `prep` step needs to be rewritten for that
platform. These are all the GHC versions for which a released version of
`weeder` exists, so it should cover all users.

Arguably, these binaries should be released by the weeder project
itself. I may open an Issue soon to see if they'd be open to upstreaming
something like tooling I have here, which is pretty non-standard due to
the need to version by GHC and not by weeder.