Build and utilize pre-compiled binaries
Making users wait for a `stack install weeder` is an unnecessary tax on
every CI run that lacks a complete cache. Instead, this PR adds an
"Binaries" action that builds binaries across a runner/GHC matrix and
stores them on a Release for use by the action. The Binaries action is
idempotent and has been run and added all the binaries already, which
our own CI uses successfully.

I made binaries for Linux and Mac and GHCs 8.10 to 9.2. Windows may be
possible, but the `prep` step needs to be rewritten for that platform.
There is no `weeder` version available for 9.4, and attempting to build
one for <= 8.8 fails with a panic. I added options to support "bring
your own `weeder`" for users where these versions don't work.

I added more `inputs` details in the `README` including a mapping of
`weeder-version` by GHC. It would be more ergonomic to allow users
specify a GHC and we just select the right weeder version. There's a few
ways to do that, so I'm deferring for now while I consider the options;
we can easily address that later.

Storing assets named by GHC version (not weeder-version) would also be
required if we wanted to ship binaries for more than one GHC per weeder
version. For example, weeder-2.2.0 should support 8.8 and 8.10, but we
only get one `weeder-2.2.0` asset, which is currently for 8.10.