Fail on any exit code if configured to fail
And don't fail regardless of exit code if not configured to fail. This
replaces the previous behavior of ignoring exit code and look for output
in the log to determine if weeds were found or not.

Before, a fatal error that produced no stdout, would not fail. This is
much better. The only gap it leaves is if we're configured _not_ to
fail, we probably still want fatal errors other than weeds-found to fail
the Job. We will resolve that next.

Fixes #13.