Update conditions around unused functions
This is behavior neutral, but it corrects that fact that we output
"unused functions found" for any error, regardless of the version of
weeder (i.e. even when we can trust its exit code to know). Now this
should output "unused functions found" or "some other error" correctly.

We did the code the way it was originally to minimize diff when
introducing the logic based on weeder version, which required larger
changes elsewhere.