feat!: rewrite in TypeScript
TypeScript actions are our preference. And with the libraries available
(`toolcache`, `exec`, etc.), there's not a ton of benefit to using the
composite action and shell script style.

I've also found the `semver-compare` action we're depending on is not
maintained, and I'd like to sunset my fork. Rewriting in TypeScript
allows using the `semver` package directly.

BREAKING CHANGE:

Instead of writing the raw weeder CLI output to the file named in the
`log` output, we now write a list of `Weed` values as JSON.